feat: upgrade all dev dependencies
BREAKING CHANGES:  is exported as a named export and no longer as the default export
