+++ .gitignore
@@ -1,5 +1,4 @@
-.*/
-coverage/
-node_modules/
-npm-debug.log
-temp/
+*.log
+.DS_Store
+node_modules
+dist
