--- .npmignore
@@ -1,8 +1,0 @@
-.*
-node_modules
-modules
-npm-debug.log
-test
-*.md
-*.config.js
-*.json
