+++ .travis.yml
@@ -1,7 +1,7 @@
 language: node_js
 cache: yarn
 node_js:
-  - '8'
+  - '12'
 script:
   - npm run lint
   - npm run build
