+++ package.json
@@ -1,68 +1,46 @@
 {
-    "name": "path-parser",
-    "version": "4.2.0",
-    "description": "A small utility to parse, match and generate paths",
-    "main": "dist/cjs/path-parser.js",
-    "jsnext:main": "dist/es/path-parser.js",
-    "module": "dist/es/path-parser.js",
-    "sideEffects": false,
-    "typings": "./typings/Path.d.ts",
-    "scripts": {
-        "test": "mocha -r ts-node/register 'test/main.js'",
-        "lint": "tslint modules/*.ts",
-        "build": "npm run clean && rollup -c rollup.config.js",
-        "clean": "rimraf dist && rimraf temp",
-        "clog": "conventional-changelog -p angular -i CHANGELOG.md -s",
-        "precommit": "lint-staged"
-    },
-    "lint-staged": {
-        "modules/*.ts": [
-            "prettier --write",
-            "tslint",
-            "git add"
-        ],
-        "test/*.js": [
-            "prettier --write",
-            "git add"
-        ]
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/troch/path-parser.git"
-    },
-    "keywords": [
-        "path",
-        "parser",
-        "url",
-        "route",
-        "match",
-        "partial match",
-        "regex"
-    ],
-    "author": "Thomas Roch",
-    "license": "ISC",
-    "bugs": {
-        "url": "https://github.com/troch/path-parser/issues"
-    },
-    "homepage": "https://github.com/troch/path-parser",
-    "devDependencies": {
-        "conventional-changelog": "~1.1.18",
-        "conventional-changelog-cli": "~2.0.1",
-        "husky": "~0.14.3",
-        "lint-staged": "~7.0.0",
-        "mkdirp": "~0.5.1",
-        "mocha": "~5.0.4",
-        "prettier": "~1.11.1",
-        "rimraf": "~2.6.2",
-        "rollup": "~0.57.1",
-        "rollup-plugin-typescript2": "~0.12.0",
-        "should": "~13.2.1",
-        "ts-node": "~5.0.1",
-        "tslint": "~5.9.1",
-        "tslint-config-prettier": "~1.10.0",
-        "typescript": "~2.7.2"
-    },
-    "dependencies": {
-        "search-params": "2.1.3"
+  "name": "path-parser",
+  "version": "4.2.0",
+  "description": "A small utility to parse, match and generate paths",
+  "license": "MIT",
+  "main": "dist/index.js",
+  "typings": "dist/index.d.ts",
+  "module": "dist/index.esm.js",
+  "sideEffects": false,
+  "files": [
+    "dist",
+    "LICENSE"
+  ],
+  "scripts": {
+    "start": "tsdx watch",
+    "build": "tsdx build",
+    "test": "tsdx test",
+    "lint": "tsdx lint",
+    "prepare": "tsdx build",
+    "clog": "conventional-changelog -p angular -i CHANGELOG.md -s"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "tsdx lint"
     }
+  },
+  "author": "Thomas Roch",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/troch/path-parser.git"
+  },
+  "bugs": {
+    "url": "https://github.com/troch/path-parser/issues"
+  },
+  "homepage": "https://github.com/troch/path-parser",
+  "dependencies": {
+    "search-params": "2.1.3"
+  },
+  "devDependencies": {
+    "@types/jest": "^24.0.25",
+    "husky": "^3.1.0",
+    "tsdx": "^0.12.1",
+    "tslib": "^1.10.0",
+    "typescript": "^3.7.4"
+  }
 }
