--- tslint.json
@@ -1,9 +1,0 @@
-{
-    "defaultSeverity": "error",
-    "extends": ["tslint:recommended", "tslint-config-prettier"],
-    "jsRules": {},
-    "rules": {
-        "object-literal-sort-keys": false
-    },
-    "rulesDirectory": []
-}
