--- typings/tokeniser.d.ts
@@ -1,9 +1,0 @@
-export interface IToken {
-    type: string;
-    match: string;
-    val: any;
-    otherVal: any;
-    regex: RegExp;
-}
-declare const tokenise: (str: string, tokens?: IToken[]) => IToken[];
-export default tokenise;
